+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## dev
 - add `VoltageRegulator`, `ShuntImpedance`, and `ShuntAdmittance` as a new bus inputs
+- add `VariableContainer = Dict{Int, Dict{String, Any}}` and `VARIABLE_NAMES`
 - add `Network.var_name_map` for parsing results
 
 ## v0.4.0